Refactor ClientLoop
