--- src/main/java/arrays/Ask.java
@@ -1,9 +1,0 @@
-package arrays;
-
-public class Ask {
-
-    public static int askType(int num) {
-        return num == 1 ? 1 : num == 2 ? 2 : num == 3 ? 3 : 0;
-    }
-
-}
