--- src/main/java/client/Main.java
@@ -1,10 +1,0 @@
-package client;
-
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-
-public class Main {
-    public static void main(String[] args) throws IOException {
-    }
-}
